Implementação de exportação de arquivo em .CSV

+++ CepAPI/CepAPI/Repository/LocalizacaoRepository.cs
@@ -2,34 +2,35 @@
 using CepAPI.Model;
 using CepAPI.MySqlContext;
 using Microsoft.EntityFrameworkCore;
+using System.Threading.Tasks;
 
 namespace CepAPI.Repository
 {
     public class LocalizacaoRepository : ILocalizacaoRepository
     {
         private readonly Context _context;
+
         public LocalizacaoRepository(Context context)
         {
             _context = context;
         }
+
         public async Task<Localizacao> CreateLocalizacao(Localizacao localizacao)
         {
             var ret = await _context.Localizacao.AddAsync(localizacao);
-
             await _context.SaveChangesAsync();
-
             ret.State = EntityState.Detached;
-
             return ret.Entity;
         }
 
         public async Task<bool> DeleteLocalAsync(int Id)
         {
             var item = await _context.Localizacao.FindAsync(Id);
+            if (item == null)
+                return false;
+
             _context.Localizacao.Remove(item);
-
             await _context.SaveChangesAsync();
-
             return true;
         }
 
@@ -42,8 +43,12 @@ namespace CepAPI.Repository
         public async Task<int> UpdateLocalizacao(Localizacao user)
         {
             _context.Entry(user).State = EntityState.Modified;
-
             return await _context.SaveChangesAsync();
+        }
+
+        public async Task<Localizacao> GetLocalizacaoById(int id)
+        {
+            return await _context.Localizacao.FirstOrDefaultAsync(l => l.Id == id);
         }
     }
 }
